+++ renderVideo.js
@@ -1,0 +1,3 @@
+import {renderVideo} from '../scripts/dist/reddit/index.js';
+
+renderVideo();
