--- scripts/.env.example
@@ -1,10 +1,0 @@
-AWS_ACCESS_KEY_ID=
-AWS_SECRET_ACCESS_KEY=
-AWS_S3_BUCKET_NAME=
-AWS_S3_REGION=
-AZURE_TTS_KEY=
-AZURE_TTS_REGION=
-REDDIT_CLIENT_ID=
-REDDIT_SECRET=
-REDDIT_USERNAME=
-REDDIT_PASSWORD=
