--- scripts/.github/workflows/render-video.yml
@@ -1,38 +1,0 @@
-name: Render video
-on:
-    workflow_dispatch:
-        inputs:
-            titleText:
-                description: "Which text should it say?"
-                required: true
-                default: "Welcome to Remotion"
-            titleColor:
-                description: "Which color should it be in?"
-                required: true
-                default: "black"
-jobs:
-    render:
-        name: Render video
-        runs-on: ubuntu-latest
-        steps:
-            - uses: actions/checkout@main
-            - uses: actions/setup-node@main
-            - run: sudo apt update
-
-            - run: sudo apt install ffmpeg
-            - run: npm i
-            - run: echo $WORKFLOW_INPUT > input-props.json
-              env:
-                  WORKFLOW_INPUT: ${{ toJson(github.event.inputs) }}
-            - run: npm run build -- --props="./input-props.json"
-              env:
-                  AZURE_TTS_KEY: ${{secrets.AZURE_TTS_KEY}}
-                  AZURE_TTS_REGION: ${{secrets.AZURE_TTS_REGION}}
-                  AWS_S3_BUCKET_NAME: ${{secrets.AWS_S3_BUCKET_NAME}}
-                  AWS_S3_REGION: ${{secrets.AWS_S3_REGION}}
-                  AWS_ACCESS_KEY_ID: ${{secrets.AWS_ACCESS_KEY_ID}}
-                  AWS_SECRET_ACCESS_KEY: ${{secrets.AWS_SECRET_ACCESS_KEY}}
-            - uses: actions/upload-artifact@v2
-              with:
-                  name: out.mp4
-                  path: out/video.mp4
