--- scripts/.gitignore
@@ -1,7 +1,0 @@
-node_modules
-dist
-.DS_Store
-
-# Ignore the output video from Git but not videos you import into src/.
-out
-.env
