--- scripts/.vscode/settings.json
@@ -1,9 +1,0 @@
-{
-	"editor.tabSize": 2,
-	"typescript.tsdk": "node_modules/typescript/lib",
-	"editor.codeActionsOnSave": {
-		"source.organizeImports": false,
-		"source.fixAll": true
-	},
-	"typescript.enablePromptUseWorkspaceTsdk": true
-}
