--- scripts/Dockerfile
@@ -1,39 +1,0 @@
-# This is a dockerized version of a server that you can easily deploy somewhere.
-# If you don't want server rendering, you can safely delete this file.
-
-FROM node:alpine
-
-# Installs latest Chromium (85) package.
-RUN apk add --no-cache \
-  chromium \
-  nss \
-  freetype \
-  freetype-dev \
-  harfbuzz \
-  ca-certificates \
-  ttf-freefont \
-  ffmpeg
-
-# Tell Puppeteer to skip installing Chrome. We'll be using the installed package.
-ENV PUPPETEER_SKIP_CHROMIUM_DOWNLOAD=true \
-  PUPPETEER_EXECUTABLE_PATH=/usr/bin/chromium-browser
-
-COPY package*.json ./
-COPY tsconfig.json ./
-COPY src src
-COPY *.ts .
-COPY *.tsx .
-
-RUN npm i
-
-# Add user so we don't need --no-sandbox.
-RUN addgroup -S pptruser && adduser -S -g pptruser pptruser \
-  && mkdir -p /home/pptruser/Downloads /app \
-  && chown -R pptruser:pptruser /home/pptruser \
-  && chown -R pptruser:pptruser /app
-# Run everything after as non-privileged user.
-USER pptruser
-
-EXPOSE 8000
-
-CMD ["npm", "run", "server"]
