--- scripts/package.json
@@ -1,43 +1,0 @@
-{
-	"name": "content-gen",
-	"version": "1.0.0",
-	"description": "My Remotion video",
-	"type": "module",
-	"scripts": {
-		"start": "remotion preview src/index.tsx --props=../scripts/posts/38c433c475fb127249f1972b38f6764c.json",
-		"build": "remotion render src/index.tsx Main out/video.mp4 --props=../scripts/posts/317ee97f491483e1f813c9f8b2f37f7c.json",
-		"upgrade": "remotion upgrade",
-		"server": "ts-node server.tsx",
-		"test": "eslint src --ext ts,tsx,js,jsx && tsc"
-	},
-	"repository": {},
-	"license": "UNLICENSED",
-	"dependencies": {
-		"@aws-sdk/client-s3": "^3.17.0",
-		"@aws-sdk/s3-request-presigner": "^3.66.0",
-		"@emotion/react": "^11.9.0",
-		"@emotion/styled": "^11.8.1",
-		"@mui/material": "^5.6.2",
-		"@remotion/bundler": "^2.3.2",
-		"@remotion/cli": "^2.3.2",
-		"@remotion/eslint-config": "^2.3.2",
-		"@remotion/renderer": "^2.3.2",
-		"@types/express": "^4.17.9",
-		"@types/react": "^17.0.0",
-		"dotenv": "^9.0.2",
-		"eslint": "^7.15.0",
-		"express": "^4.17.1",
-		"md5": "^2.3.0",
-		"microsoft-cognitiveservices-speech-sdk": "^1.17.0",
-		"prettier": "^2.2.1",
-		"react": "^17.0.2",
-		"react-dom": "^17.0.2",
-		"remotion": "^2.3.2",
-		"ts-node": "^10.0.0",
-		"typescript": "4.4.2"
-	},
-	"devDependencies": {
-		"@types/md5": "^2.3.0",
-		"@types/node-fetch": "^2.6.1"
-	}
-}
