--- scripts/remotion.config.ts
@@ -1,4 +1,0 @@
-import {Config} from 'remotion';
-
-Config.Rendering.setImageFormat('jpeg');
-Config.Output.setOverwriteOutput(true);
