--- scripts/renderVideo.js
@@ -1,3 +1,0 @@
-import {renderVideo} from '../../scripts/dist/findPosts.js';
-
-renderVideo();
