--- scripts/src/index.tsx
@@ -1,4 +1,0 @@
-import {registerRoot} from 'remotion';
-import {RemotionVideo} from './Video';
-
-registerRoot(RemotionVideo);
