--- scripts/testRedditApi.js
@@ -1,3 +1,0 @@
-import {getRedditData} from '../../scripts/dist/findPosts.js';
-
-getRedditData();
