--- scripts/tsconfig.json
@@ -1,14 +1,0 @@
-{
-	"compilerOptions": {
-		"target": "ES2018",
-		"module": "commonjs",
-		"jsx": "react-jsx",
-		"outDir": "./dist",
-		"strict": true,
-		"noEmit": true,
-		"lib": ["es2022", "DOM"],
-		"esModuleInterop": true,
-		"skipLibCheck": true,
-		"forceConsistentCasingInFileNames": true
-	}
-}
