+++ src/Video.tsx
@@ -1,35 +1,54 @@
 import {useState, useEffect, useCallback} from 'react';
-import {getAudioDuration} from '@remotion/media-utils';
+import {getVideoMetadata} from '@remotion/media-utils';
 import {
 	Composition,
 	continueRender,
 	delayRender,
 	getInputProps,
 } from 'remotion';
-import {textToSpeech} from './TextToSpeech';
-import {HelloWorld} from './HelloWorld';
+import {Main} from './Main';
+import {
+	createIntro,
+	createBody,
+	createBodyFromComments,
+	calculateDuration,
+	calculateSegmentDuration,
+} from './utilities';
+import {ContentSegments} from './types';
 
 const inputProps = getInputProps();
-
+const defaultContentValue = {
+	intro: undefined,
+	body: undefined,
+};
 export const RemotionVideo: React.FC = () => {
 	const [handle] = useState(() => delayRender());
-	const [duration, setDuration] = useState<number>(300);
+	const [videoFrames, setVideoFrames] = useState<number>(1);
+	const [totalFrames, setTotalFrames] = useState<number>(1);
+	const [content, setContent] =
+		useState<ContentSegments>(defaultContentValue);
 
-	const getDuration = useCallback(async () => {
-		console.log(inputProps);
-		const url = await textToSpeech(
-			inputProps.title + ' ' + inputProps.titleText,
-			'enUSWoman2'
+	const initVideoData = useCallback(async () => {
+		const data = {intro: {}, body: [{}]};
+		data.intro = await createIntro(inputProps.post.title);
+		const body = await createBody(inputProps.post, inputProps.user.data);
+		data.body = body.concat(
+			await createBodyFromComments(
+				inputProps.comments.postComments,
+				inputProps.comments.users
+			)
 		);
-		const imported = await getAudioDuration(url);
-		const duration = Math.round(imported) * 30;
-		setDuration(duration);
+		const duration = await getVideoMetadata(inputProps.video);
+		const videoFrames = Math.round(duration.durationInSeconds) * 30;
+		setVideoFrames(videoFrames);
+		setTotalFrames(calculateDuration(data as ContentSegments));
+		setContent(calculateSegmentDuration(data as ContentSegments));
 		continueRender(handle);
 	}, [handle]);
 
 	useEffect(() => {
-		getDuration();
-	}, [getDuration, handle]);
+		initVideoData();
+	}, [initVideoData]);
 
 	if (!process.env.AZURE_TTS_KEY) {
 		throw new Error(
@@ -64,15 +83,15 @@ export const RemotionVideo: React.FC = () => {
 	return (
 		<>
 			<Composition
-				id="HelloWorld"
-				component={HelloWorld}
-				durationInFrames={duration}
+				id="Main"
+				component={Main}
+				durationInFrames={totalFrames}
 				fps={30}
-				width={1920}
-				height={1080}
+				width={1080}
+				height={1920}
 				defaultProps={{
-					titleText: '',
-					titleColor: 'black',
+					videoFrames,
+					content,
 				}}
 			/>
 		</>
