+++ tsconfig.json
@@ -6,7 +6,7 @@
 		"outDir": "./dist",
 		"strict": true,
 		"noEmit": true,
-		"lib": ["es2015", "DOM"],
+		"lib": ["ES2021", "DOM"],
 		"esModuleInterop": true,
 		"skipLibCheck": true,
 		"forceConsistentCasingInFileNames": true
